coding

+++ workspace/Study/src/BAEKJOON/가장긴증가하는부분수열4_14002.java
@@ -5,22 +5,22 @@ public class 가장긴증가하는부분수열4_14002 {
 	public static void main(String[] args) {
 		Scanner sc = new Scanner(System.in);
 		int N = sc.nextInt();
-		int[] arr = new int[N];
+		long[] arr = new long[N];
 
 		for (int i = 0; i < N; i++)
-			arr[i] = sc.nextInt();
+			arr[i] = sc.nextLong();
 
 		
-		ArrayList<Integer> []d = new ArrayList[N];
+		ArrayList<Long> []d = new ArrayList[N];
 		for(int i=0; i<N; i++) {
-			d[i] = new ArrayList<Integer>();
+			d[i] = new ArrayList<Long>();
 		}
 
 		int result = 1;
 		int resultx= 0;
 		for (int i = 0; i < N; i++) {
 			
-			int max = 0;
+			long max = 0;
 			int buf = -1;
 			for (int j = 0; j < i; j++) {
 				if (arr[j] < arr[i] && max < d[j].size()) {
